requirements.txt
aiogram==3.21.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.1
bcrypt==4.2.0
reportlab==4.2.5
qrcode==8.0
openpyxl==3.1.5
